+++ packages/@withstudiocms/internal_helpers/test/utils/loadChangelog.test.ts
@@ -1,6 +1,6 @@
 /** biome-ignore-all lint/suspicious/noExplicitAny: allowed for tests */
 import { describe, expect, it } from 'vitest';
-import { loadChangelog, semverCategories } from '../../src/utils/index.js';
+import { loadChangelog, semverCategories } from '../../src/index.js';
 
 // Example changelog markdown for testing
 const exampleMarkdown = `
